Update OPCClient.cpp
I was having issues with using the code in a multi-threaded environment.  I ended up discovering that the init() and stop() functions were changed in a way that was causing issues in a multi-threaded environment where the CoGetMalloc was throwing an exception due to iMalloc already being initialized (which occurred in a different thread).  This update protects the iMalloc memory so it doesn't get reinitialized on subsequent calls to init().  The purpose of this commit is to restore the initialization and stop functions to the way they were implemented in the original version.